requirements.txt
fastapi
uvicorn
python-dotenv
numpy
requests
openai
resemblyzer
openai-whisper
